+++ ui/src/util.ts
@@ -9,20 +9,11 @@ export const isSequenceColumn = (column: PColumnIdAndSpec) => {
     column.domain?.['pl7.app/alphabet'] === 'aminoacid';
   const isSingleCellSequence = (column: PColumnSpec) =>
     column.domain?.['pl7.app/vdj/scClonotypeChain/index'] === 'primary'
+    && column.domain?.['pl7.app/alphabet'] === 'aminoacid'
     // && column.axesSpec.length >= 1
     && column.axesSpec[0].name === 'pl7.app/vdj/scClonotypeKey';
 
   return isBulkSequence(column.spec) || isSingleCellSequence(column.spec);
-};
-
-export const isLabelColumnOption = (column: PColumnIdAndSpec) => {
-  if (column.spec.axesSpec.length === 1 && column.spec.axesSpec[0].name === 'pl7.app/sampleId')
-    return false;
-  return true;
-};
-
-export const isLinkerColumn = (column: PColumnIdAndSpec) => {
-  return column.spec.annotations?.['pl7.app/isLinkerColumn'] === 'true';
 };
 
 export function defaultFilters(tSpec: PTableColumnSpec): (PlTableFilter | undefined) {