+++ block/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @platforma-open/milaboratories.clonotype-space
 
+## 2.2.5
+
+### Patch Changes
+
+- Updated dependencies [520fdef]
+  - @platforma-open/milaboratories.clonotype-space.ui@2.4.5
+
+## 2.2.4
+
+### Patch Changes
+
+- Updated dependencies [7379529]
+- Updated dependencies [ed663b8]
+  - @platforma-open/milaboratories.clonotype-space.ui@2.4.4
+
 ## 2.2.3
 
 ### Patch Changes