+++ ui/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @platforma-open/milaboratories.clonotype-space.ui
 
+## 2.4.5
+
+### Patch Changes
+
+- 520fdef: update graph-maker version
+
+## 2.4.4
+
+### Patch Changes
+
+- 7379529: MSA updates
+- ed663b8: Only allow aa alignment
+
 ## 2.4.3
 
 ### Patch Changes